src/components/file_uploader.py
import streamlit as st
import os

def file_uploader():
    st.title("Medical Report Chatbot")
    st.write("Upload your medical report (PDF format) to get interpretations.")

    uploaded_file = st.file_uploader("Choose a PDF file", type="pdf")

    if uploaded_file is not None:
        # Save the uploaded file to the uploads directory
        uploads_dir = os.path.join("data", "uploads")
        os.makedirs(uploads_dir, exist_ok=True)
        file_path = os.path.join(uploads_dir, uploaded_file.name)

        with open(file_path, "wb") as f:
            f.write(uploaded_file.getbuffer())

        st.success("File uploaded successfully!")

        # Here you can add functionality to process the uploaded file
        # For example, call the processing function and display results
        # interpretation = process_uploaded_file(file_path)
        # st.write(interpretation)