pyproject.toml

[tool.poetry]
package-mode = false

[project]
name = "medical-report-copilot"
version = "0.1.0"
description = "This package is for the project using a multimodal LLM to interprete medical reports and images"
authors = [
    {name = "Jeanette"}
]
license = {text = "To be determined"}
readme = "README.md"
requires-python = ">=3.12, <4.0"
dependencies = [
    "openai (>=1.83.0,<2.0.0)",
    "litellm (>=1.72.0,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pydantic (>=2.11.5,<3.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "python-json-logger (>=3.3.0,<4.0.0)",
    "hydra-core (>=1.3.2,<2.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "pypdf (>=5.6.0,<6.0.0)",
    "instructor (>=1.8.3,<2.0.0)",
    "pymupdf (>=1.26.0,<2.0.0)",
    "traceloop-sdk (>=0.40.7,<0.41.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
